LPS-126770 Moving to DDM service (no logic changes)
It is not safe to use SQL for tables managed by other modules. It is
safer to place this upgrade in DDM. In the next commit we'll remove the
SQL that references DLFileEntry.
