+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/upgrade/DDMServiceUpgrade.java
@@ -53,6 +53,7 @@ import com.liferay.dynamic.data.mapping.internal.upgrade.v3_5_0.UpgradeDDMFormIn
 import com.liferay.dynamic.data.mapping.internal.upgrade.v3_7_1.UpgradeDDMStructureEmptyValidation;
 import com.liferay.dynamic.data.mapping.internal.upgrade.v4_0_0.UpgradeDDMField;
 import com.liferay.dynamic.data.mapping.internal.upgrade.v4_3_0.UpgradeDLFileEntryTypeDDMFieldAttribute;
+import com.liferay.dynamic.data.mapping.internal.upgrade.v4_3_0.UpgradeDLFileEntryTypeDataDefinitionId;
 import com.liferay.dynamic.data.mapping.io.DDMFormDeserializer;
 import com.liferay.dynamic.data.mapping.io.DDMFormLayoutDeserializer;
 import com.liferay.dynamic.data.mapping.io.DDMFormLayoutSerializer;
@@ -399,7 +400,8 @@ public class DDMServiceUpgrade implements UpgradeStepRegistrator {
 				UpgradeDDMFormInstanceRecord());
 
 		registry.register(
-			"4.2.0", "4.3.0", new UpgradeDLFileEntryTypeDDMFieldAttribute());
+			"4.2.0", "4.3.0", new UpgradeDLFileEntryTypeDDMFieldAttribute(),
+			new UpgradeDLFileEntryTypeDataDefinitionId());
 	}
 
 	@Activate
